+++ reduction/tests/test_dead_time.py
@@ -1,7 +1,7 @@
 import mantid.simpleapi as mtd_api
 
 from lr_reduction import template
-from lr_reduction.DeadTimeCorrection import SingleReadoutDeadTimeCorrection
+from lr_reduction.dead_time_correction import SingleReadoutDeadTimeCorrection
 from lr_reduction.utils import amend_config
 
 mtd_api.config["default.facility"] = "SNS"
@@ -48,9 +48,9 @@ def test_deadtime_paralyzable(nexus_dir):
         assert c > 0
         assert c < 1.001
 
-def test_deadtime_threshold_ratio(nexus_dir):
+def test_deadtime_threshold(nexus_dir):
     """
-    Test using the threshold ratio. Here the threshold is set to 0,
+    Test using the threshold. Here the threshold is set to 0,
     so all corrections will be 0
     """
     with amend_config(data_dir=nexus_dir):
@@ -60,12 +60,12 @@ def test_deadtime_threshold_ratio(nexus_dir):
     algo.PyInit()
     algo.setProperty("InputWorkspace", ws)
     algo.setProperty("OutputWorkspace", "dead_time_corr")
-    algo.setProperty("ThresholdRatio", 0.0)
+    algo.setProperty("DeadTimeThreshold", 1.1)
     algo.PyExec()
     corr_ws = algo.getProperty("OutputWorkspace").value
     corr = corr_ws.readY(0)
     for c in corr:
-        assert c == 0
+        assert c <= 1.1
 
 def test_full_reduction(nexus_dir):
     """
@@ -88,28 +88,3 @@ def test_full_reduction(nexus_dir):
     for c in corr:
         assert c > 0
         assert c < 1.001
-
-
-def test_full_reduction_with_threshold(nexus_dir):
-    """
-    Test dead time from the reduction workflow using a
-    threshold ratio. Here the threshold is zero,
-    so all corrections will be set to zero
-    """
-    template_path = "data/template.xml"
-    with amend_config(data_dir=nexus_dir):
-        ws = mtd_api.Load("REF_L_198409")
-
-    sequence_number = ws.getRun().getProperty("sequence_number").value[0]
-    template_data = template.read_template(template_path, sequence_number)
-    template_data.dead_time = True
-    template_data.dead_time_threshold_ratio = 0.5
-
-    _, r1, _ = template.process_from_template_ws(ws, template_data)
-
-    template_data.dead_time = False
-    _, r2, _ = template.process_from_template_ws(ws, template_data)
-
-    corr = r1 / r2
-    for c in corr:
-        assert c == 0.5